+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Test to validate the requirement of zipcode for billing address.
 
 ## [0.5.0] - 2021-07-30
+
 ### Added
--  Test for invoice address without zipcode.
+
+- Test for invoice address without zipcode.
 
 ## [0.4.9] - 2021-07-29